Dart multi-platform builds through conditional imports and exports.
* Reorganize dart:io usages and support 'io', 'html'.
  * 'ui' and 'isolate' TBD.
* Rewrite bit operations with 32-bit version because Dart2JS does not support 64-bit integers.
