+++ runtime/Dart/lib/src/error/error.dart
@@ -4,6 +4,9 @@
  * can be found in the LICENSE.txt file in the project root.
  */
 
+export 'src/console_error_listener_stub.dart'
+    if (dart.library.io) 'src/console_error_listener_io.dart'
+    if (dart.library.html) 'src/console_error_listener_html.dart';
 export 'src/diagnostic_error_listener.dart';
 export 'src/error_listener.dart';
 export 'src/error_strategy.dart';
