+++ runtime/Dart/lib/src/error/src/error_listener.dart
@@ -4,8 +4,6 @@
  * can be found in the LICENSE.txt file in the project root.
  */
 
-import 'dart:io';
-
 import '../../atn/atn.dart';
 import '../../dfa/dfa.dart';
 import '../../parser.dart';
@@ -219,26 +217,6 @@ class BaseErrorListener extends ErrorListener {
     String msg,
     RecognitionException? e,
   ) {}
-}
-
-class ConsoleErrorListener extends BaseErrorListener {
-  /// Provides a default instance of [ConsoleErrorListener].
-  static final INSTANCE = ConsoleErrorListener();
-
-  /// {@inheritDoc}
-  ///
-  /// <p>
-  /// This implementation prints messages to {@link System//err} containing the
-  /// values of [line], [charPositionInLine], and [msg] using
-  /// the following format.</p>
-  ///
-  /// <pre>
-  /// line <em>line</em>:<em>charPositionInLine</em> <em>msg</em>
-  /// </pre>
-  @override
-  void syntaxError(recognizer, offendingSymbol, line, column, msg, e) {
-    stderr.writeln('line $line:$column $msg');
-  }
 }
 
 /// This implementation of [ErrorListener] dispatches all calls to a
