+++ runtime/Dart/lib/src/util/file_io.dart
@@ -1,0 +1,3 @@
+import 'dart:io';
+
+Stream<List<int>> readStream(String path) => File(path).openRead();
