+++ runtime/Dart/lib/src/util/file_stub.dart
@@ -1,0 +1,2 @@
+Stream<List<int>> readStream(String path) =>
+    throw UnsupportedError('Cannot read a file stream without dart:io');
