+++ runtime/Dart/lib/src/util/write_html.dart
@@ -1,0 +1,3 @@
+import 'dart:html';
+
+void write(Object? object) => window.console.info(object);
