+++ runtime/Dart/lib/src/util/write_io.dart
@@ -1,0 +1,3 @@
+import 'dart:io';
+
+void write(Object? object) => stdout.write(object);
