+++ runtime/Dart/lib/src/util/write_stub.dart
@@ -1,0 +1,1 @@
+void write(Object? object) => print(object);
